rename precompiles
