+++ CMakeLists.txt
@@ -135,16 +135,16 @@ add_library(
   "include/monad/execution/evmone_baseline_interpreter.hpp"
   "include/monad/execution/ethereum/dao.hpp"
   "include/monad/execution/ethereum/genesis.hpp"
-  "include/monad/execution/ethereum/static_precompiles/elliptic_curve_recover.hpp"
-  "include/monad/execution/ethereum/static_precompiles/sha256_hash.hpp"
-  "include/monad/execution/ethereum/static_precompiles/ripemd160_hash.hpp"
-  "include/monad/execution/ethereum/static_precompiles/identity.hpp"
-  "include/monad/execution/ethereum/static_precompiles/modular_exponentiation.hpp"
-  "include/monad/execution/ethereum/static_precompiles/bn_add.hpp"
-  "include/monad/execution/ethereum/static_precompiles/bn_multiply.hpp"
-  "include/monad/execution/ethereum/static_precompiles/bn_pairing.hpp"
-  "include/monad/execution/ethereum/static_precompiles/blake2f.hpp"
-  "include/monad/execution/static_precompiles.hpp"
+  "include/monad/execution/precompiles.hpp"
+  "include/monad/execution/precompiles/elliptic_curve_recover.hpp"
+  "include/monad/execution/precompiles/sha256_hash.hpp"
+  "include/monad/execution/precompiles/ripemd160_hash.hpp"
+  "include/monad/execution/precompiles/identity.hpp"
+  "include/monad/execution/precompiles/modular_exponentiation.hpp"
+  "include/monad/execution/precompiles/bn_add.hpp"
+  "include/monad/execution/precompiles/bn_multiply.hpp"
+  "include/monad/execution/precompiles/bn_pairing.hpp"
+  "include/monad/execution/precompiles/blake2f.hpp"
   "include/monad/execution/execution_model.hpp"
   "include/monad/execution/replay_block_db.hpp"
   "include/monad/execution/transaction_processor.hpp"
